+++ frontend/src/JumpToOffset.svelte
@@ -42,7 +42,7 @@
 
   $: if (mounted) {
     startOffset = Math.max(
-      Math.ceil((dataLength * $scrollY.top) / alignment) * alignment,
+      Math.floor((dataLength * $scrollY.top) / alignment) * alignment,
       0
     );
     input.value = `0x${startOffset.toString(16)}`;
@@ -55,7 +55,7 @@
     if (e.key === 'Enter') {
       input.blur();
       try {
-        let result = calculator.calculate(input.value);
+        let result = calculator.calculate(input.value) + 1;
         $scrollY.top = result / dataLength;
       } catch (_) {
         input.value = `0x${startOffset.toString(16)}`;