+++ frontend/src/ofrak/resource.js
@@ -382,6 +382,14 @@ export class Resource {
     throw new NotImplementedError("search_for_vaddr");
   }
 
+  async update_script() {
+    throw new NotImplementedError("update_script");
+  }
+
+  async add_flush_to_disk_to_script() {
+    throw new NotImplementedError("add_flush_to_disk_to_script");
+  }
+
   /***
    * Return the string representation of a comment, which includes its range as prefix if it
    * has one.