+++ frontend/src/stores.js
@@ -5,3 +5,10 @@ import { writable } from "svelte/store";
  */
 export const selected = writable(undefined);
 export const selectedResource = writable(undefined);
+export const script = writable([]);
+
+/***
+ * Points the OFRAK frontend to a seperate backend server
+ * When empty, uses the same host and port as the frontend
+ */
+export const backendUrl = "";