+++ ofrak_core/ofrak/model/component_filters.py
@@ -85,7 +85,7 @@ class AnalyzerOutputFilter(ComponentFilter):
         def component_is_analyzer_with_outputs(c: ComponentInterface):
             if _isinstance(c, Analyzer):
                 analyzer_outputs = set(c.get_outputs_as_attribute_types())  # type: ignore
-                return not analyzer_outputs.isdisjoint(self.outputs)
+                return self.outputs.issubset(analyzer_outputs)
             return False
 
         return {c for c in components if component_is_analyzer_with_outputs(c)}