+++ ofrak_core/test_ofrak/components/test_uf2_component.py
@@ -18,11 +18,17 @@ FILENAME = "rp2-pico-20220618-v1.19.1.uf2"
 EXPECTED_DATA = b"Raspberry Pi Pico with RP1337"
 
 
+async def test_uf2_identify(ofrak_context: OFRAKContext) -> None:
+    asset_path = Path(test_ofrak.components.ASSETS_DIR, FILENAME)
+    root_resource = await ofrak_context.create_root_resource_from_file(str(asset_path))
+    await root_resource.identify()
+    assert root_resource.has_tag(Uf2File), "Expected resource to have tag Uf2File"
+
+
 class TestUf2UnpackModifyPack(UnpackModifyPackPattern):
     async def create_root_resource(self, ofrak_context: OFRAKContext) -> Resource:
         asset_path = Path(test_ofrak.components.ASSETS_DIR, FILENAME)
         root_resource = await ofrak_context.create_root_resource_from_file(str(asset_path))
-        root_resource.add_tag(Uf2File)
         await root_resource.save()
         return root_resource
 
@@ -53,4 +59,3 @@ class TestUf2UnpackModifyPack(UnpackModifyPackPattern):
         resource_data = await repacked_uf2_resource.get_data()
         unpacked_data = resource_data[0x78EB5:0x78ED2]
         assert unpacked_data == EXPECTED_DATA
-        assert repacked_uf2_resource.has_tag(Uf2File)